Пятый коммит. 1.Добавлена проверка на подачу данных. Если данные не поданы выходит сообщение. 2. Результат деления теперь может быть дробным.

+++ src/main/java/pro/sky/calculator/CalculatorService.java
@@ -5,5 +5,5 @@ public interface CalculatorService {
     int sum(int num1, int num2);
     int difference(int num1, int num2);
     int multiply(int num1, int num2);
-    int divide(int num1, int num2);
+    double divide(int num1, int num2);
 }
