+++ src/main/java/pro/sky/calculator/CalculatorServiceImpl.java
@@ -25,9 +25,9 @@ public class CalculatorServiceImpl implements CalculatorService{
     }
 
     @Override
-    public int divide(int num1, int num2) {
+    public double divide(int num1, int num2) {
         if (num2 != 0) {
-            return num1 / num2;
+            return (double) num1 / num2;
         } else
             return 0;
     }
